Fixed some small issues. Unique notifications request should be good.
Client's request to get unique notifications when going from home page to notification should be good to go. The request is being made to the server side where the JSON file is resided. The server takes the request and handles it in the virtualDatabase which returns the wanted JSON object back to the server and the server returns the JSON object back to the client side.
The process is depicted as the following:
prototype.js -> server.js -> virtualDatabase.js -> server.js -> prototype.js